mod: 댓글, 게시글 좋아요 기능 수정

+++ src/main/java/com/project/foradhd/domain/board/business/service/Impl/GeneralCommentServiceImpl.java
@@ -1,12 +1,12 @@
 package com.project.foradhd.domain.board.business.service.Impl;
 
+import com.project.foradhd.domain.board.business.service.CommentLikeService;
 import com.project.foradhd.domain.board.business.service.GeneralCommentService;
 import com.project.foradhd.domain.board.persistence.entity.GeneralComment;
-import com.project.foradhd.domain.board.persistence.entity.GeneralPost;
 import com.project.foradhd.domain.board.persistence.repository.GeneralCommentRepository;
 import com.project.foradhd.domain.board.web.dto.GeneralCommentDto;
-import com.project.foradhd.domain.board.web.dto.GeneralPostDto;
 import com.project.foradhd.domain.board.web.mapper.GeneralCommentMapper;
+import com.project.foradhd.global.exception.CommentNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
@@ -20,17 +20,19 @@ public class GeneralCommentServiceImpl implements GeneralCommentService {
 
     private final GeneralCommentRepository repository;
     private final GeneralCommentMapper mapper;
+    private final CommentLikeService commentLikeService;
 
     @Autowired
-    public GeneralCommentServiceImpl(GeneralCommentRepository repository, GeneralCommentMapper mapper) {
+    public GeneralCommentServiceImpl(GeneralCommentRepository repository, GeneralCommentMapper mapper, CommentLikeService commentLikeService) {
         this.repository = repository;
         this.mapper = mapper;
+        this.commentLikeService = commentLikeService;
     }
 
     @Override
     public GeneralCommentDto getComment(String commentId) {
         GeneralComment comment = repository.findById(commentId)
-                .orElseThrow(() -> new RuntimeException("Comment not found"));
+                .orElseThrow(() -> new CommentNotFoundException("Comment not found with ID: " + commentId));
         return mapper.toDto(comment);
     }
 
@@ -43,13 +45,17 @@ public class GeneralCommentServiceImpl implements GeneralCommentService {
 
     @Override
     public void deleteComment(String commentId) {
+        if (!repository.existsById(commentId)) {
+            throw new CommentNotFoundException("Comment not found with ID: " + commentId);
+        }
         repository.deleteById(commentId);
     }
+
 
     @Override
     public GeneralCommentDto updateComment(GeneralCommentDto commentDto) {
         GeneralComment existingComment = repository.findById(commentDto.getCommentId())
-                .orElseThrow(() -> new RuntimeException("Comment not found"));
+                .orElseThrow(() -> new CommentNotFoundException("Comment not found with ID: " + commentDto.getCommentId()));
         existingComment = mapper.toEntity(commentDto);
         repository.save(existingComment);
         return mapper.toDto(existingComment);
@@ -62,5 +68,10 @@ public class GeneralCommentServiceImpl implements GeneralCommentService {
         return comments.stream()
                 .map(mapper::toDto)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    public void toggleCommentLike(String userId, String commentId) {
+        commentLikeService.toggleLike(userId, commentId);
     }
 }
